requirements.txt
attrs==21.2.0
Django==2.2
djangorestframework==3.9.2
importlib-metadata==4.8.1
iniconfig==1.1.1
model-mommy==2.0.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
zipp==3.6.0
